requirements.txt
supabase
dotenv
langchain
pymupdf==1.25.1
streamlit
langchain-google-genai
pyjwt
